Moved shared rest handler code into RestHandlerUtils
