+++ src/main/kotlin/com/amazon/opendistroforelasticsearch/indexstatemanagement/resthandler/RestAddPolicyAction.kt
@@ -18,6 +18,9 @@ package com.amazon.opendistroforelasticsearch.indexstatemanagement.resthandler
 import com.amazon.opendistroforelasticsearch.indexstatemanagement.IndexStateManagementPlugin.Companion.ISM_BASE_URI
 import com.amazon.opendistroforelasticsearch.indexstatemanagement.elasticapi.getPolicyName
 import com.amazon.opendistroforelasticsearch.indexstatemanagement.settings.ManagedIndexSettings
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.util.FailedIndex
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.util.UPDATED_INDICES
+import com.amazon.opendistroforelasticsearch.indexstatemanagement.util.buildInvalidIndexResponse
 import org.elasticsearch.action.admin.cluster.state.ClusterStateRequest
 import org.elasticsearch.action.admin.cluster.state.ClusterStateResponse
 import org.elasticsearch.action.admin.indices.settings.put.UpdateSettingsAction
@@ -30,7 +33,6 @@ import org.elasticsearch.cluster.block.ClusterBlockException
 import org.elasticsearch.cluster.metadata.IndexMetaData
 import org.elasticsearch.common.Strings
 import org.elasticsearch.common.settings.Settings
-import org.elasticsearch.common.xcontent.XContentBuilder
 import org.elasticsearch.common.xcontent.XContentHelper
 import org.elasticsearch.index.Index
 import org.elasticsearch.rest.BaseRestHandler
@@ -120,7 +122,7 @@ class RestAddPolicyAction(settings: Settings, controller: RestController) : Base
                                     })
                                 }
 
-                                buildInvalidIndexResponse(builder)
+                                buildInvalidIndexResponse(builder, failedIndices)
                                 return BytesRestResponse(RestStatus.OK, builder.endObject())
                             }
                         }
@@ -133,29 +135,12 @@ class RestAddPolicyAction(settings: Settings, controller: RestController) : Base
                             "Failed to add policy due to ClusterBlockingException: ${e.message}")
                     })
 
-                    buildInvalidIndexResponse(builder)
+                    buildInvalidIndexResponse(builder, failedIndices)
                     channel.sendResponse(BytesRestResponse(RestStatus.OK, builder.endObject()))
                 }
             } else {
-                buildInvalidIndexResponse(builder)
+                buildInvalidIndexResponse(builder, failedIndices)
                 channel.sendResponse(BytesRestResponse(RestStatus.OK, builder.endObject()))
-            }
-        }
-
-        private fun buildInvalidIndexResponse(builder: XContentBuilder) {
-            if (failedIndices.isNotEmpty()) {
-                builder.field(FAILURES, true)
-                builder.startArray(FAILED_INDICES)
-                for (failedIndex in failedIndices) {
-                    builder.startObject()
-                    builder.field("index_name", failedIndex.name)
-                    builder.field("index_uuid", failedIndex.uuid)
-                    builder.field("reason", failedIndex.reason)
-                    builder.endObject()
-                }
-                builder.endArray()
-            } else {
-                builder.field(FAILURES, false)
             }
         }
 
@@ -185,12 +170,7 @@ class RestAddPolicyAction(settings: Settings, controller: RestController) : Base
         }
     }
 
-    data class FailedIndex(val name: String, val uuid: String, val reason: String)
-
     companion object {
         const val ADD_POLICY_BASE_URI = "$ISM_BASE_URI/add"
-        const val FAILURES = "failures"
-        const val FAILED_INDICES = "failed_indices"
-        const val UPDATED_INDICES = "updated_indices"
     }
 }
